chore(wrapped-keys): LIT-3877 - Remove lit actions (and their generated counterparts) from wrapped-keys package - Defined initial stub of `wrapped-keys-lit-actions` because git got confused about deleting vs. moving files
