+++ packages/wrapped-keys-lit-actions/buffer.shim.js
@@ -1,0 +1,2 @@
+import { Buffer } from 'buffer';
+globalThis.Buffer = Buffer;
