+++ packages/wrapped-keys-lit-actions/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "@lit-protocol/wrapped-keys-lit-actions",
+  "type": "commonjs",
+  "license": "MIT",
+  "homepage": "https://github.com/Lit-Protocol/js-sdk",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/LIT-Protocol/js-sdk"
+  },
+  "keywords": [
+    "library"
+  ],
+  "bugs": {
+    "url": "https://github.com/LIT-Protocol/js-sdk/issues"
+  },
+  "publishConfig": {
+    "access": "public",
+    "directory": "../../dist/packages/wrapped-keys-lit-actions"
+  },
+  "tags": [
+    "universal"
+  ],
+  "buildOptions": {
+    "genReact": false
+  },
+  "version": "6.5.0",
+  "main": "./dist/undefined",
+  "typings": "./dist/src/index.d.ts"
+}
