+++ packages/wrapped-keys-lit-actions/src/index.ts
@@ -1,0 +1,3 @@
+import generateEncryptedEthereumPrivateKey from './lib/ethereum/generateEncryptedEthereumPrivateKey';
+
+export { generateEncryptedEthereumPrivateKey };
