+++ packages/wrapped-keys-lit-actions/src/lib/ethereum/signTransactionWithEthereumEncryptedKey.js
@@ -1,4 +1,4 @@
-const { removeSaltFromDecryptedKey } = require('../../utils');
+const { removeSaltFromDecryptedKey } = require('../utils');
 
 /**
  *
